py3k: Move http client to a file with a different name to unmask the import on py3k
The name needs to be different from the package name so the test can import the
correct name.
