+++ src/collectors/http/http_collector.py
